所有id欄位的類別從int改為long

+++ InsuranceSystem/src/main/java/com/project01/controller/InsuranceCompanyController.java
@@ -59,7 +59,7 @@ public class InsuranceCompanyController {
 	}
 	
 	@DeleteMapping(path = "/delete/{id}",produces = "application/json")
-	public ResponseEntity<ResponseDTO<InsuranceCompany>> deleteInsuranceCompanyController(@PathVariable("id") int companyId) {
+	public ResponseEntity<ResponseDTO<InsuranceCompany>> deleteInsuranceCompanyController(@PathVariable("id") long companyId) {
 		return  insuranceCompanyService.deleteInsuranceCompany(companyId);
 	}
